% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_input_node.R
\name{add_input_node}
\alias{add_input_node}
\title{Add an input node to a dot graph}
\usage{
add_input_node(
  x,
  node = "raw_data",
  ignore = "(get|build|make)",
  col = "#000000",
  style = "rectangle"
)
}
\arguments{
\item{x}{dot file as a character vector}

\item{node}{The name of the node to add}

\item{ignore}{Regular expression used to define any orphan nodes which
should not be added to}

\item{col}{The colour for the input node}

\item{style}{The shape of the node}
}
\value{
A character vector
}
\description{
Add an additional input node to a dot graph
}
\details{
Takes a dot file as a character vector, such as would be obtained
using readLines, and adds a node to any existing nodes with no parental node.
Orphan nodes which should be ignored can be specified using the pattern in
the \code{ignore} argument.
}
