% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_node_colour.R
\name{change_node_colour}
\alias{change_node_colour}
\title{Change the colour of a given node}
\usage{
change_node_colour(x, pattern, col)
}
\arguments{
\item{x}{dot file as a character vector}

\item{pattern}{Regular expression used to define one or more nodes}

\item{col}{The new node colour}
}
\value{
A character vector
}
\description{
Change the colour of any nodes matching a pattern
}
\details{
Finds nodes matching the regular expression provided and sets their
colour as provided
}
