% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corner.R
\name{corner}
\alias{corner}
\alias{corner.matrix}
\alias{corner.data.frame}
\title{Display one corner of a large matrix-like object}
\usage{
corner(x, ...)

\method{corner}{matrix}(
  x,
  corner = c("topleft", "topright", "bottomleft", "bottomright"),
  n = 6L,
  ...
)

\method{corner}{data.frame}(
  x,
  corner = c("topleft", "topright", "bottomleft", "bottomright"),
  n = 6L,
  ...
)
}
\arguments{
\item{x}{A matrix or data.frame}

\item{...}{arguments to be passed to or from other methods.}

\item{corner}{The corner to show}

\item{n}{The number of rows and columns to show}
}
\description{
Display one corner of a large matrix-like object
}
\details{
This provides an alternative to head for matrix-like objects with many
columns
}
\section{Methods (by class)}{
\itemize{
\item \code{matrix}: Display one corner of a large matrix-like object

\item \code{data.frame}: Display one corner of a large matrix-like object
}}

