% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotExpressionPCA.R
\name{plotExpressionPCA}
\alias{plotExpressionPCA}
\alias{plotExpressionPCA,DGEList-method}
\alias{plotExpressionPCA,SummarizedExperiment-method}
\title{Create a PCA plot from a DGEList object}
\usage{
plotExpressionPCA(object, ...)

\S4method{plotExpressionPCA}{DGEList}(object, ..., .center = TRUE, .scale = FALSE)

\S4method{plotExpressionPCA}{SummarizedExperiment}(object, ..., .center = TRUE, .scale = FALSE)
}
\arguments{
\item{object}{An object containing expression values (i.e. counts)}

\item{...}{Used for passing parameters to autoplot.prcomp. These should be
column names and can be passed to plotting aesthetics such as colour, shape
and size.}

\item{.center}{passed to \code{prcomp}}

\item{.scale}{passed to \code{prcomp}}
}
\value{
A ggplot2 object
}
\description{
Plot the PCA results from a DGEList
}
\examples{
library(edgeR)
y <- matrix(rnbinom(10000,mu=5,size=2), ncol=4)
d <- DGEList(counts=y, group=rep(1:2, each=2))
plotExpressionPCA(d, colour = "group")

library(SummarizedExperiment)
nrows <- 200; ncols <- 6
counts <- matrix(runif(nrows * ncols, 1, 1e4), nrows)
colData <- DataFrame(Treatment=rep(c("ChIP", "Input"), 3),
                     row.names=LETTERS[1:6])
se0 <- SummarizedExperiment(assays=SimpleList(counts=counts),
                            colData=colData)
plotExpressionPCA(se0, colour = "Treatment")

}
