% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_dot_node.R
\name{rm_dot_node}
\alias{rm_dot_node}
\title{Remove a single node from a snakemake dag}
\usage{
rm_dot_node(x, node = "all")
}
\arguments{
\item{x}{dot file as a character vector}

\item{node}{The name of node to drop}
}
\value{
A character vector
}
\description{
Remove a single node from a snakemake dag
}
\details{
Takes a dot file as a character vector, such as would be obtained
using readLines, and removes a single node.
Commonly this is the 'all' node as might be produced from a snakemake
rulegraph

The resultant graph can then be visualised using \code{grViz} from the
package DiagrammeR, or passed into any function which takes dot format as a
character vector
}
