% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voomWeightsFromCPM.R
\name{voomWeightsFromCPM}
\alias{voomWeightsFromCPM}
\title{Estimate Voom Precision Weights Directly From CPM Values}
\usage{
voomWeightsFromCPM(
  cpm,
  design = NULL,
  w0 = NULL,
  lib.size = NULL,
  isLogCPM = TRUE,
  span = 0.5,
  ...
)
}
\arguments{
\item{cpm}{Matrix of CPM or logCPM values}

\item{design}{The design matrix for the experiment}

\item{w0}{Initial vector of sample weights. Should be calculated using
\code{\link{arrayWeights}}}

\item{lib.size}{Initial library sizes. Must be provided as these are no
estimable from CPM values}

\item{isLogCPM}{logical(1). Indicates whether the data is log2 transformed
already. Most commonly (e.g. if using the output of cqn) it will be,}

\item{span}{Width of the smoothing window used for the lowess mean-variance
trend. Expressed as a proportion between 0 and 1.}

\item{...}{Passed to lmFit internally}
}
\value{
A list with elements \code{Amean}, \code{sigma} and \code{w}.
The weights for downstream analysis are in the element \code{w}, whilst
the values \code{Amean} and \code{sigma} may be useful for plotting
the mean-variance relationship before incorporation of the precision
weights.
Typically, you would plot \code{sqrt{sigma}} on the y-axis to replicate
the default plots of the original \code{voom} function

If sample weights were provided, modified weights will also be returned, as
the initial function \code{\link{voomWithQualityWeights}} performs two
rounds of estimation of sample weights.
Here we would simply provide the initial weights round a priori, with the
second round performed within the function.
Importantly, this second round of sample weight estimation uses the precision
weights ensuring the correct mean-variance relationship is used for the
estimation of sample weights
}
\description{
Estimate Voom Precision Weights Directly From CPM Values
}
\details{
This function takes CPM or logCPM values and estimates the precision weights
as would be done by providing counts directly to the \code{\link{voom}}
function.
Using this function enables the use of logCPM values which have been
normalised using other methods such as Conditional-Quantile or
Smooth-Quantile Normalisation.

The precision weights are returned as part of the output, and these are able
to be passed to the function \code{\link{lmFit}} during model fitting.
This will ensure that the mean-variance relationship is appropriate for
the linear modelling steps as performed by limma.

An implementation incorporating sample weights is yet to written
}
